requirements.txt
streamlit
selenium
webdriver_manager.chrome
pandas
requests
